# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DogMap API",
    "description": "API REST para sistema de locais pet-friendly.\n\n## Sobre o Projeto\nO DogMap é uma plataforma que permite aos usuários descobrir, avaliar e favoritar locais que aceitam pets.\n\n## Funcionalidades Principais\n- **Usuários**: Cadastro e gerenciamento de usuários com informações de seus pets\n- **Locais**: Catálogo de estabelecimentos pet-friendly com diferentes categorias\n- **Avaliações**: Sistema de avaliações e comentários para os locais\n- **Favoritos**: Lista personalizada de locais favoritos por usuário\n\n## Tecnologias\n- Node.js + Express\n- Dados mockados para desenvolvimento\n- Swagger para documentação",
    "version": "1.0.0",
    "contact": {
      "name": "DogMap Team",
      "email": "contato@dogmap.com",
      "url": "https://dogmap.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3009",
      "description": "Servidor de Desenvolvimento"
    },
    {
      "url": "https://api.dogmap.com",
      "description": "Servidor de Produção"
    }
  ],
...
  "paths": {
...
    "/api/places": {
      "get": {
        "summary": "Lista todos os locais",
        "description": "Retorna uma lista de locais pet-friendly com possibilidade de filtro por tipo",
        "tags": ["Locais"],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filtrar locais por tipo específico",
            "schema": {
              "type": "string",
              "enum": ["Pet Shop", "Clínica Veterinária", "Parque", "Shopping", "Praia", "Restaurante", "Hotel"]
            },
            "example": "Pet Shop"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de locais retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Place"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Número total de locais",
                      "example": 10
                    },
                    "filters": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "description": "Tipo de filtro aplicado",
                          "example": "Pet Shop"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Cadastra um novo local",
        "description": "Cria um novo local pet-friendly no sistema",
        "tags": ["Locais"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "type"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nome do estabelecimento",
                    "example": "Pet Shop Amigo Fiel",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "type": {
                    "type": "string",
                    "description": "Categoria do local",
                    "enum": ["Pet Shop", "Clínica Veterinária", "Parque", "Shopping", "Praia", "Restaurante", "Hotel"],
                    "example": "Pet Shop"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Local criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Local criado com sucesso"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Place"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/places/{id}": {
      "get": {
        "summary": "Busca um local por ID",
        "description": "Retorna as informações detalhadas de um local específico",
        "tags": ["Locais"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único do local",
            "schema": {
              "type": "string"
            },
            "example": "place_456"
          }
        ],
        "responses": {
          "200": {
            "description": "Local encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Place"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Place": {
        "type": "object",
        "description": "Informações do local pet-friendly",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do local",
            "example": "place_456"
          },
          "name": {
            "type": "string",
            "description": "Nome do estabelecimento",
            "example": "Pet Shop Amigo Fiel"
          },
          "type": {
            "type": "string",
            "description": "Categoria do local",
            "example": "Pet Shop",
            "enum": ["Pet Shop", "Clínica Veterinária", "Parque", "Shopping", "Praia", "Restaurante", "Hotel"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação do local",
            "example": "2024-01-15T10:30:00Z"
          }
        },
        "required": ["id", "name", "type"]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Dados inválidos ou obrigatórios não fornecidos",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Dados obrigatórios não fornecidos"
                },
                "message": {
                  "type": "string",
                  "example": "Nome, email e informações do pet são obrigatórios"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Usuário não encontrado"
                },
                "message": {
                  "type": "string",
                  "example": "Usuário com ID user_123 não foi encontrado"
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                },
                "message": {
                  "type": "string",
                  "example": "Ocorreu um erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT para autenticação (implementação futura)"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}